Add files via upload
Not very good yet, but it's a start. Issues:

the first time it never sounds the advance sound. it does not seem to save the progress. i never want it to ask the same question twice in a row. the color picker has gone missing. in level 7 the questions (and answers) go above 100. the favicon is missing